rename growi to slack
