+++ src/server/routes/apiv3/index.js
@@ -46,7 +46,7 @@ module.exports = (crowi) => {
   router.use('/bookmarks', require('./bookmarks')(crowi));
   router.use('/attachment', require('./attachment')(crowi));
 
-  router.use('/growi-bot', require('./growi-bot')(crowi));
+  router.use('/slack-bot', require('./slack-bot')(crowi));
 
   return router;
 };
